server/requirements.txt
flask==3.0.3
flask-cors
flask-dance==7.1.0
flask_login
python-decouple==3.4
python-dotenv
gunicorn==20.1.0
pyopenssl===21.0.0
Werkzeug==3.0.3
redis